+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 * 1.0.0-next (@release_date@)
 
  * added `NullDiscovery`
+ * changed `ResourceDiscovery::find()` and `getBindings()` to throw a
+   `NoSuchTypeException` if the type has not been defined
 
 * 1.0.0-beta (2015-01-12)